extracted inventory drop handler and inventory event handler
