+++ handlers/inspect_item_handler.py
@@ -1,6 +1,7 @@
 from engine import Engine
 from entity import Item
-from handlers.input_handlers import EventHandler, InventoryEventHandler
+from handlers.inventory_event_handler import InventoryEventHandler
+from handlers.input_handlers import EventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
