+++ handlers/inventory_binds_handler.py
@@ -1,6 +1,6 @@
 import color
 from entity import Item
-from handlers.input_handlers import InventoryEventHandler
+from handlers.inventory_event_handler import InventoryEventHandler
 from handlers.input_handlers import ActionOrHandler
 from typing import Optional
